requirements.txt
# Core ML / PyTorch
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
pytorch-lightning==2.1.0

# Hugging Face ecosystem
transformers==4.27.4
accelerate==0.20.3
safetensors==0.5.3
huggingface-hub

# stable_audio_tools dependencies (pinned because stable_audio_tools is installed from source)
# Install stable_audio_tools via: pip install git+https://github.com/Stability-AI/stable-audio-tools.git
einops==0.8.1
ema-pytorch==0.2.3
encodec==0.1.1
descript-audio-codec==1.0.0
k-diffusion==0.1.1
v-diffusion-pytorch==0.0.2
vector-quantize-pytorch==1.14.41
x-transformers==1.26.6
prefigure

# Audio processing
librosa==0.11.0
soundfile==0.13.1
audioread==3.0.1

# Computer vision & image processing
opencv-python
mediapipe
Pillow
# Install OpenAI CLIP via: pip install git+https://github.com/openai/CLIP.git

# Data & computation
numpy
scipy
matplotlib
tqdm

# HTTP & environment
requests
python-dotenv

# Experiment tracking (used via pytorch-lightning and prefigure)
wandb==0.15.4
tensorboard==2.19.0

# Web interface
fastapi==0.116.0
uvicorn==0.35.0
celery

# RunPod serverless
runpod
